+++ src/app/components/crew-area/crew-area.component.ts
@@ -14,7 +14,6 @@ import { MatInputModule } from '@angular/material/input';
 import { MatTableDataSource, MatTableModule } from '@angular/material/table';
 import { MatSort, MatSortModule } from '@angular/material/sort';
 import { MatPaginatorModule } from '@angular/material/paginator';
-import { WcV3Service } from '@services/api/wc-v3.service';
 import { DateFormatPipe } from '@pipes/date-format.pipe';
 import { DISCLAIMER_PRODUCTS } from '@models/cross-sale.model';
 import { CommonModule } from '@angular/common';
@@ -23,6 +22,15 @@ import { CustomEndpointsService } from '@services/api/custom-endpoints.service';
 import { MatSelectModule } from '@angular/material/select';
 import { LineItemMin, OrderMin } from '@models/types.model';
 import { WC_ORDER_STATUSES, WcOrderStatus } from '@models/order.model';
+import { MatIconModule } from '@angular/material/icon';
+import { MatMenuModule } from '@angular/material/menu';
+import { WcV3Service } from '@services/api/wc-v3.service';
+import { SelectionModel } from '@angular/cdk/collections';
+import { MatCheckboxModule } from '@angular/material/checkbox';
+import { catchError, Subject, throwError } from 'rxjs';
+import { ErrorDialogService } from '@shared/errors/error-dialog.service';
+import { indicate } from '@utils/reactive-loading.utils';
+import { MatProgressSpinnerModule } from '@angular/material/progress-spinner';
 
 @Component({
   selector: 'app-crew-area',
@@ -40,6 +48,12 @@ import { WC_ORDER_STATUSES, WcOrderStatus } from '@models/order.model';
     OrderStatusComponent,
     MatSelectModule,
     MatSort,
+    MatInputModule,
+    MatIconModule,
+    MatButtonModule,
+    MatMenuModule,
+    MatCheckboxModule,
+    MatProgressSpinnerModule,
   ],
   templateUrl: './crew-area.component.html',
   styleUrl: './crew-area.component.scss',
@@ -47,7 +61,7 @@ import { WC_ORDER_STATUSES, WcOrderStatus } from '@models/order.model';
   changeDetection: ChangeDetectionStrategy.OnPush,
 })
 export class CrewAreaComponent implements AfterViewInit, OnInit {
-  endpoints = inject(WcV3Service);
+  wcV3 = inject(WcV3Service);
   customEpS = inject(CustomEndpointsService);
   contactPersons = signal<Array<string>>([]);
   dataSource = new MatTableDataSource<OrderMin>([]);
@@ -62,25 +76,23 @@ export class CrewAreaComponent implements AfterViewInit, OnInit {
     'items_cross',
     'total_price',
     'order_status',
+    'select',
   ];
+  selection = new SelectionModel<OrderMin>(true, []);
+  loading$ = new Subject<boolean>();
 
   statuses = WC_ORDER_STATUSES;
   shiftFilter = '';
   nameFilter = '';
   contactFilter: string | null = null;
-  statusFilter: WcOrderStatus | null = null;
+  statusFilter: WcOrderStatus[] = this.statuses.filter(
+    (status) => !new Set(['on-hold', 'refunded', 'failed']).has(status),
+  );
 
-  constructor() {}
+  constructor(private readonly errorService: ErrorDialogService) {}
 
   ngOnInit(): void {
-    //TODO: This can be improved
-    this.customEpS.getOrdersByInvite('', ['2025']).subscribe((orders) => {
-      this.dataSource.data = orders;
-    });
-
-    this.customEpS.listAllowedInvites().subscribe((response: string[]) => {
-      this.contactPersons.set(response);
-    });
+    this.refreshCollection();
 
     /** filter predicate */
     this.dataSource.filterPredicate = (order: OrderMin, filter: string) => {
@@ -94,7 +106,8 @@ export class CrewAreaComponent implements AfterViewInit, OnInit {
 
       const statusMatch =
         !parsedFilter.order_status ||
-        order.status === parsedFilter.order_status;
+        parsedFilter.order_status.length === 0 ||
+        parsedFilter.order_status.includes(order.status);
 
       const shiftMatch =
         !parsedFilter.items_main ||
@@ -106,6 +119,35 @@ export class CrewAreaComponent implements AfterViewInit, OnInit {
 
       return shiftMatch && contactMatch && statusMatch && nameMatch;
     };
+  }
+
+  private refreshCollection() {
+    this.loading$.next(true);
+    //TODO: This can be improved
+    this.customEpS
+      .getOrdersByInvite('', ['2025'])
+      .pipe(
+        indicate(this.loading$),
+        catchError((error) => {
+          this.errorService.handleError(error);
+          return throwError(() => error);
+        }),
+      )
+      .subscribe((orders) => {
+        this.dataSource.data = orders;
+      });
+
+    this.customEpS
+      .listAllowedInvites()
+      .pipe(
+        catchError((error) => {
+          this.errorService.handleError(error);
+          return throwError(() => error);
+        }),
+      )
+      .subscribe((response: string[]) => {
+        this.contactPersons.set(response);
+      });
   }
 
   findMainItem(items: Array<LineItemMin>) {
@@ -122,9 +164,58 @@ export class CrewAreaComponent implements AfterViewInit, OnInit {
       name: this.nameFilter.trim().toLowerCase(),
       items_main: this.shiftFilter.trim().toLowerCase(),
       invited_by: this.contactFilter,
-      order_status: this.statusFilter,
+      order_status: this.statusFilter.length ? this.statusFilter : null,
     };
     this.dataSource.filter = JSON.stringify(filterObj);
+  }
+
+  toggleRow(order: OrderMin) {
+    this.selection.toggle(order);
+  }
+
+  isAllSelected() {
+    return this.selection.selected.length === this.dataSource.data.length;
+  }
+
+  isPartialSelection() {
+    return this.selection.selected.length > 0 && !this.isAllSelected();
+  }
+
+  // toggle all rows
+  toggleAllRows() {
+    if (this.isAllSelected()) {
+      this.selection.clear();
+    } else {
+      this.dataSource.data.forEach((order: OrderMin) =>
+        this.selection.select(order),
+      );
+    }
+  }
+
+  bulkUpdateStatus(status: WcOrderStatus) {
+    const selectedOrders = this.selection.selected.map((order) => order.id);
+    this.wcV3
+      .batchUpdateOrderStatus(selectedOrders, status)
+      .pipe(
+        catchError((error) => {
+          this.errorService.handleError(error);
+          return throwError(() => error);
+        }),
+      )
+      .subscribe(() => {
+        this.refreshCollection();
+        this.selection.clear(); // reset selection
+      });
+  }
+
+  selectAll(event: Event) {
+    console.log(event);
+    if (!event.bubbles) {
+      this.statusFilter = []; // Unselect all if already all are selected
+    } else {
+      this.statusFilter = [...this.statuses]; // Select all if not already selected
+    }
+    this.applyFilter(); // Apply the filter after the change
   }
 
   ngAfterViewInit(): void {