+++ CHANGELOG.md
@@ -2,6 +2,23 @@
 
 All notable changes to this project will be documented in this file.
 
+## [1.0.7]
+
+### Added
+
+- crew-page
+
+## [1.0.6]
+
+### Changed
+
+- styling of nav
+
+### Removed
+
+- unused services and redundant code
+- non-generic dialog components
+
 ## [1.0.5]
 
 ### Added